+++ internal/models/user.go
@@ -3,7 +3,7 @@ package models
 type GetBalanceRequest string
 
 type User struct {
-	Id            int64   `json:"id"`
-	Balance       float64 `json:"balance"`
-	FrozenBalance float64 `json:"frozen_balance"`
+	Id            int64 `json:"id"`
+	Balance       int64 `json:"balance"`
+	FrozenBalance int64 `json:"frozen_balance"`
 }
