+++ scripts/create_tables.sql
@@ -2,9 +2,9 @@ CREATE TABLE IF NOT EXISTS users (
     -- unique id, not serial because of task text
     id int PRIMARY KEY,
     -- current active balance of user
-    balance money NOT NULL default 0::money CHECK(balance >= 0::money),
+    balance bigint NOT NULL default 0 CHECK(balance >= 0),
     -- reserved balance of user
-    frozen_balance money DEFAULT 0::money CHECK(frozen_balance >= 0::money)
+    frozen_balance bigint DEFAULT 0 CHECK(frozen_balance >= 0)
 );
 
 CREATE TABLE IF NOT EXISTS services (
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS orders (
     proofed_at timestamp DEFAULT null,
     user_id integer REFERENCES users(id) ON DELETE CASCADE,
     service_id INTEGER REFERENCES services(id) ON DELETE CASCADE,
-    amount money NOT NULL CHECK(amount >= 0::money),
+    amount bigint NOT NULL CHECK(amount >= 0),
     status text DEFAULT 'in_progress' CHECK(status in ('in_progress', 'success', 'declined'))
 );
